Add ability to append in any order
Implement markers to add ability to append data in any order.

Every time something is appended, the list of markers is
traversed and the last one relative to the newly appended
data is marked with the index to the newly added plus its size minus
the one word of the marker.
These two data are stored in one uint32_t, using bitwise operations,
where the highest order 16 bits are the index and the lowest are
the number of words.

Update sample with more insert before and after function calls to show
the correct functioning.