Updt CMklists & test_0 w/ array, macro file path
Update the CMakeLists with a macro definition for the unit tests;
set the macro to the full path to the assets folder so that it doesn't
have to be manually defnied in the code.

Update the test_0 source file accordingly and also change from
vectors to arrays for the instructions to inject.